+++ src/api/src/main/java/com/bankofz/inqcust/api/service/RandomCustomerNumberGenerator.java
@@ -2,5 +2,5 @@ package com.bankofz.inqcust.api.service;
 
 public interface RandomCustomerNumberGenerator {
 
-    String nextCustomerNumber(int highestCustomerNumber);
+    String nextCustomerNumber(long highestCustomerNumber);
 }
